feat: refactor all non-typer components into another file
